[jQAssistant/jqa-bom#1] Introduction of a BOM for shared dependencies
Dependencies, mainly intended for plugin development, shared by
multiple components of jQAssistant, have been managed until
now independently by many components. The goal of the introduction
of a Bill of Materials is to have a central place to define
the default scope and version of all these dependencies in a
central place.